+++ src/pages/commits/ActiveCommits/utils.ts
@@ -1,75 +1,6 @@
 import { mapTaskStatusToUmbrellaStatus } from "constants/task";
 import { ColorCount } from "pages/commits/types";
-import { ChartTypes, Commits, BuildVariantDict } from "types/commits";
-import {
-  TASK_ICONS_PER_ROW,
-  TASK_ICON_HEIGHT,
-  TASK_ICON_PADDING,
-  GROUPED_BADGES_PER_ROW,
-  GROUPED_BADGE_HEIGHT,
-  GROUPED_BADGE_PADDING,
-} from "../constants";
-
-const constructBuildVariantDict = (versions: Commits): BuildVariantDict => {
-  const buildVariantDict: BuildVariantDict = {};
-
-  for (let i = 0; i < versions.length; i++) {
-    const { version } = versions[i];
-
-    // skip if inactive/unmatching
-    if (version) {
-      // Deduplicate build variants and build variant stats by consolidating into a single object.
-      const allBuildVariants = [
-        ...version.buildVariants,
-        ...version.buildVariantStats,
-      ].reduce((acc, curr) => {
-        const { variant } = curr;
-        acc[variant] = { ...acc[variant], ...curr };
-        return acc;
-      }, {});
-
-      // Construct build variant dict which will contain information needed for rendering.
-      Object.values(allBuildVariants).reduce(
-        (acc, { tasks, statusCounts, variant }) => {
-          // Determine height to allocate for icons.
-          let iconHeight = 0;
-          if (tasks) {
-            const numRows = Math.ceil(tasks.length / TASK_ICONS_PER_ROW);
-            const iconContainerHeight = numRows * TASK_ICON_HEIGHT;
-            const iconContainerPadding = TASK_ICON_PADDING * 2;
-            iconHeight = iconContainerHeight + iconContainerPadding;
-          }
-
-          // Determine height to allocate for grouped badges.
-          let badgeHeight = 0;
-          if (statusCounts) {
-            const numRows = Math.ceil(
-              statusCounts.length / GROUPED_BADGES_PER_ROW
-            );
-            const badgeContainerHeight = numRows * GROUPED_BADGE_HEIGHT;
-            const badgeContainerPadding = GROUPED_BADGE_PADDING * 2;
-            badgeHeight = badgeContainerHeight + badgeContainerPadding;
-          }
-
-          if (acc[variant]) {
-            if (iconHeight > acc[variant].iconHeight) {
-              acc[variant].iconHeight = iconHeight;
-            }
-            if (badgeHeight > acc[variant].badgeHeight) {
-              acc[variant].badgeHeight = badgeHeight;
-            }
-            acc[variant].priority += 1;
-          } else {
-            acc[variant] = { priority: 1, iconHeight, badgeHeight };
-          }
-          return acc;
-        },
-        buildVariantDict
-      );
-    }
-  }
-  return buildVariantDict;
-};
+import { ChartTypes } from "types/commits";
 
 /**
  * `calculateBarHeight` calculates the height of a single bar in a bar chart.
@@ -166,7 +97,6 @@ const injectGlobalHighlightStyle = (taskIdentifier: string) => {
 
 export {
   calculateBarHeight,
-  constructBuildVariantDict,
   getStatusesWithZeroCount,
   injectGlobalDimStyle,
   injectGlobalHighlightStyle,
